comments and removing unused code

+++ src/game/scoreboard.h
@@ -9,15 +9,19 @@
 
 class ScoreBoard {
     private:
-        int whiteScore;
-        int blackScore;
+        //track each player's score
+        float whiteScore;
+        float blackScore;
     public:
         ScoreBoard();
         ~ScoreBoard();
 
+        //update score based on game result
         void updateScore(bool isWhiteWinner);
+        //display score at the end of the program
         void displayScore() const;
 
+        //update score based on stalemate
         void stalemateUpdate();
 };
